fix(promql): make the native-histogram result path decode, end to end (#2037)
* fix(promql): pin native-histogram result column types so rate() decodes

The nine Histogram*Column outputs took their ClickHouse type from
whichever lowering built the projection: a bare selector and sum()
forwarded the OTel-CH columns verbatim (UInt64 / Array(UInt64)), while
rate()/increase() divided by the window and landed on Float64 /
Array(Float64). chclient.HistogramValue binds one set of scan
destinations by position, so the two shapes cannot both decode --
clickhouse-go/v2's Float64.ScanRow rejects a *uint64 destination
outright, which meant rate() over an exponential histogram failed at the
first row in production.

Pin the nine output types in chsql's emitter instead: counts and sums to
Float64, bucket ladders to Array(Float64), Scale and the two bucket
offsets to Int32. Float64 is the correct pin rather than the convenient
one -- rate() over a native histogram yields fractional counts, exactly
as reference Prometheus returns a *histogram.FloatHistogram.

Refs #1967

* feat(chclient): decode native histograms on the columnar path

The columnar matrix decoder bound four columns and returned
errMatrixMismatch for the thirteen-column histogram projection. A
decline is not a slow path: columnarDecoder.decode responds by
re-dispatching the identical query to ClickHouse under a fresh
query_id, so every histogram-valued query executed twice.

Bind the nine Histogram*Column outputs too. The binder asserts one
concrete ch-go column type per slot, which is sound only because the
emitter now pins those nine output types.

Also teach internal/chclienttest the same thirteen-column shape: its
four-destination scan meant every chDB-backed handler test was
structurally unable to serve a histogram-valued query.

Tests:
- histogram_scan_contract_test.go drives clickhouse-go's own ScanRow
  for each pinned column type into the real HistogramValue destination,
  with a negative control pinning that a Float64 column genuinely
  rejects the *uint64 destination this repo used to bind. This is the
  assertion the shipped rate() lowering would have failed.
- columnar_histogram_test.go pins the widened binder: names and types
  are still checked, the four-column shape still binds without a
  histogram, and rows decode with per-row HistogramValues.
- handler_chdb_histogram_valued_test.go serves a bare selector, sum(),
  sum by (), and rate() through the real handler over chDB and asserts
  the Prometheus wire JSON against exp_histogram_rate.txtar's
  independently hand-worked calculation.

Refs #1967

---------

+++ internal/api/prom/histogram_wire_test.go
@@ -21,7 +21,7 @@ func TestHistogramFromValue_PositiveBucketBoundaries(t *testing.T) {
 		Sum:                  10,
 		Scale:                0, // base = 2^(2^0) = 2
 		PositiveOffset:       1,
-		PositiveBucketCounts: []uint64{2, 3},
+		PositiveBucketCounts: []float64{2, 3},
 	}
 	h := histogramFromValue(hv)
 	if h == nil {
@@ -57,7 +57,7 @@ func TestHistogramFromValue_NegativeBucketBoundariesAscending(t *testing.T) {
 	hv := &chclient.HistogramValue{
 		Scale:                0, // base = 2
 		NegativeOffset:       0,
-		NegativeBucketCounts: []uint64{5, 7}, // i=0 -> [-2,-1); i=1 -> [-4,-2)
+		NegativeBucketCounts: []float64{5, 7}, // i=0 -> [-2,-1); i=1 -> [-4,-2)
 	}
 	h := histogramFromValue(hv)
 	if len(h.Buckets) != 2 {
@@ -82,8 +82,8 @@ func TestHistogramFromValue_ZeroBucketSymmetric(t *testing.T) {
 	hv := &chclient.HistogramValue{
 		ZeroThreshold:        0.5,
 		ZeroCount:            9,
-		PositiveBucketCounts: []uint64{1},
-		NegativeBucketCounts: []uint64{1},
+		PositiveBucketCounts: []float64{1},
+		NegativeBucketCounts: []float64{1},
 	}
 	h := histogramFromValue(hv)
 	var zero *HistogramBucket
@@ -113,7 +113,7 @@ func TestHistogramFromValue_ZeroBucketOneSidedClamp(t *testing.T) {
 		hv := &chclient.HistogramValue{
 			ZeroThreshold:        0.5,
 			ZeroCount:            4,
-			PositiveBucketCounts: []uint64{1},
+			PositiveBucketCounts: []float64{1},
 		}
 		h := histogramFromValue(hv)
 		zero := findBucketByCount(t, h.Buckets, "4")
@@ -126,7 +126,7 @@ func TestHistogramFromValue_ZeroBucketOneSidedClamp(t *testing.T) {
 		hv := &chclient.HistogramValue{
 			ZeroThreshold:        0.5,
 			ZeroCount:            4,
-			NegativeBucketCounts: []uint64{1},
+			NegativeBucketCounts: []float64{1},
 		}
 		h := histogramFromValue(hv)
 		zero := findBucketByCount(t, h.Buckets, "4")
@@ -156,8 +156,8 @@ func TestHistogramFromValue_SkipsZeroCountBuckets(t *testing.T) {
 	t.Parallel()
 
 	hv := &chclient.HistogramValue{
-		PositiveBucketCounts: []uint64{0, 5, 0},
-		NegativeBucketCounts: []uint64{0},
+		PositiveBucketCounts: []float64{0, 5, 0},
+		NegativeBucketCounts: []float64{0},
 		ZeroCount:            0,
 	}
 	h := histogramFromValue(hv)
@@ -298,7 +298,7 @@ func TestToVector_PopulatesHistogram(t *testing.T) {
 	t.Parallel()
 
 	ts := time.Unix(1717171717, 0).UTC()
-	hv := &chclient.HistogramValue{Count: 3, Sum: 6, PositiveBucketCounts: []uint64{3}}
+	hv := &chclient.HistogramValue{Count: 3, Sum: 6, PositiveBucketCounts: []float64{3}}
 	samples := []chclient.Sample{
 		{MetricName: "h", Labels: map[string]string{"job": "api"}, Timestamp: ts, Histogram: hv},
 	}
@@ -327,7 +327,7 @@ func TestMatrixFromSamples_PopulatesHistogram(t *testing.T) {
 	t.Parallel()
 
 	ts := time.Unix(1717171717, 0).UTC()
-	hv := &chclient.HistogramValue{Count: 1, PositiveBucketCounts: []uint64{1}}
+	hv := &chclient.HistogramValue{Count: 1, PositiveBucketCounts: []float64{1}}
 	samples := []chclient.Sample{
 		{MetricName: "h", Labels: map[string]string{"job": "api"}, Timestamp: ts, Histogram: hv},
 	}
@@ -351,7 +351,7 @@ func TestMatrixFromCursor_PopulatesHistogram(t *testing.T) {
 	t.Parallel()
 
 	ts := time.Unix(1717171717, 0).UTC()
-	hv := &chclient.HistogramValue{Count: 1, PositiveBucketCounts: []uint64{1}}
+	hv := &chclient.HistogramValue{Count: 1, PositiveBucketCounts: []float64{1}}
 	samples := []chclient.Sample{
 		{MetricName: "h", Labels: map[string]string{"job": "api"}, Timestamp: ts, Histogram: hv},
 	}
